% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{calaminer-package}
\alias{calaminer}
\alias{calaminer-package}
\title{calaminer: Fast Excel File Reader Using Rust}
\description{
High-performance Excel file reader supporting xlsx, xlsm, xlsb, xls, and ods formats. Built on the Rust 'calamine' library, providing significantly faster read times compared to pure R implementations. Supports reading sheet names, dimensions, and data as data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MaxManek1110/calamine_r}
  \item Report bugs at \url{https://github.com/MaxManek1110/calamine_r/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Manek \email{maximilian.manek@hotmail.de}

Authors:
\itemize{
  \item Johann Tuffe (calamine Rust library)
  \item Authors of the dependent Rust crates (see AUTHORS file)
}

}
\keyword{internal}
