% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calamine.R
\name{excel_sheets}
\alias{excel_sheets}
\title{Get Sheet Names from Excel File}
\usage{
excel_sheets(path)
}
\arguments{
\item{path}{Path to the Excel file}
}
\value{
Character vector of sheet names
}
\description{
Get Sheet Names from Excel File
}
\examples{
test_file <- system.file("extdata", "test.xlsx", package = "calaminer")
if (nzchar(test_file)) {
  sheets <- excel_sheets(test_file)
  print(sheets)
}
}
